requirements.txt
streamlit
google-generativeai
psycopg2-binary
python-dotenv
markdown
flask
gunicorn
sentence-transformers
pgvector
scikit-learn
numpy
requests